+++ previous-versions/accounting/src/sdk/sdk.ts
@@ -77,9 +77,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.40.1";
-    genVersion = "2.143.2";
-    userAgent = "speakeasy-sdk/typescript 0.40.1 2.143.2 3.0.0 @codat/accounting";
+    sdkVersion = "0.40.2";
+    genVersion = "2.144.7";
+    userAgent = "speakeasy-sdk/typescript 0.40.2 2.144.7 3.0.0 @codat/accounting";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);