+++ previous-versions/accounting/RELEASES.md
@@ -739,3 +739,13 @@ Based on:
 - [typescript v0.40.1] previous-versions/accounting
 ### Releases
 - [NPM v0.40.1] https://www.npmjs.com/package/@codat/accounting/v/0.40.1 - previous-versions/accounting
+
+## 2023-10-03 00:03:50
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.40.2] previous-versions/accounting
+### Releases
+- [NPM v0.40.2] https://www.npmjs.com/package/@codat/accounting/v/0.40.2 - previous-versions/accounting